Fixed bugs in registration page, login page and dashboard page
